Fixed more bugs | Added requiredPermission option | DEV BUILD
